+++ libs/build_script/src/rust/main/lib.rs
@@ -13,4 +13,4 @@
 
 pub mod console;
 pub mod tailwind;
-
+pub mod resources;